+++ edc-c-client/protobuf/src/google/protobuf/descriptor.cc
@@ -131,7 +131,7 @@ string ToCamelCase(const string& input) {
   string result;
   result.reserve(input.size());
 
-  for (int i = 0; i < input.size(); i++) {
+  for (int i = 0; i < (int)(input.size()); i++) {
     if (input[i] == '_') {
       capitalize_next = true;
     } else if (capitalize_next) {
@@ -489,7 +489,7 @@ DescriptorPool::Tables::~Tables() {
   // Note that the deletion order is important, since the destructors of some
   // messages may refer to objects in allocations_.
   STLDeleteElements(&messages_);
-  for (int i = 0; i < allocations_.size(); i++) {
+  for (int i = 0; i < (int)(allocations_.size()); i++) {
     operator delete(allocations_[i]);
   }
   STLDeleteElements(&strings_);
@@ -513,13 +513,13 @@ void DescriptorPool::Tables::Checkpoint() {
 }
 
 void DescriptorPool::Tables::Rollback() {
-  for (int i = 0; i < symbols_after_checkpoint_.size(); i++) {
+  for (int i = 0; i < (int)(symbols_after_checkpoint_.size()); i++) {
     symbols_by_name_.erase(symbols_after_checkpoint_[i]);
   }
-  for (int i = 0; i < files_after_checkpoint_.size(); i++) {
+  for (int i = 0; i < (int)(files_after_checkpoint_.size()); i++) {
     files_by_name_.erase(files_after_checkpoint_[i]);
   }
-  for (int i = 0; i < extensions_after_checkpoint_.size(); i++) {
+  for (int i = 0; i < (int)(extensions_after_checkpoint_.size()); i++) {
     extensions_.erase(extensions_after_checkpoint_[i]);
   }
 
@@ -533,7 +533,7 @@ void DescriptorPool::Tables::Rollback() {
     messages_.begin() + messages_before_checkpoint_, messages_.end());
   STLDeleteContainerPointers(
     file_tables_.begin() + file_tables_before_checkpoint_, file_tables_.end());
-  for (int i = allocations_before_checkpoint_; i < allocations_.size(); i++) {
+  for (int i = allocations_before_checkpoint_; i < (int)(allocations_.size()); i++) {
     operator delete(allocations_[i]);
   }
 
@@ -986,7 +986,7 @@ void DescriptorPool::FindAllExtensions(
     vector<int> numbers;
     if (fallback_database_->FindAllExtensionNumbers(extendee->full_name(),
                                                     &numbers)) {
-      for (int i = 0; i < numbers.size(); ++i) {
+      for (int i = 0; i < (int)(numbers.size()); ++i) {
         int number = numbers[i];
         if (tables_->FindExtension(extendee, number) == NULL) {
           TryFindExtensionInFallbackDatabase(extendee, number);
@@ -1495,7 +1495,7 @@ bool RetrieveOptions(const Message &options, vector<string> *option_entries) {
   const Reflection* reflection = options.GetReflection();
   vector<const FieldDescriptor*> fields;
   reflection->ListFields(options, &fields);
-  for (int i = 0; i < fields.size(); i++) {
+  for (int i = 0; i < (int)(fields.size()); i++) {
     // Doesn't make sense to have message type fields here
     if (fields[i]->cpp_type() == FieldDescriptor::CPPTYPE_MESSAGE) {
       continue;
@@ -1531,7 +1531,7 @@ bool FormatLineOptions(int depth, const Message &options, string *output) {
   string prefix(depth * 2, ' ');
   vector<string> all_options;
   if (RetrieveOptions(options, &all_options)) {
-    for (int i = 0; i < all_options.size(); i++) {
+    for (int i = 0; i < (int)(all_options.size()); i++) {
       strings::SubstituteAndAppend(output, "$0option $1;\n",
                                    prefix, all_options[i]);
     }
@@ -2604,7 +2604,7 @@ void DescriptorBuilder::ValidateSymbolName(
     AddError(full_name, proto, DescriptorPool::ErrorCollector::NAME,
              "Missing name.");
   } else {
-    for (int i = 0; i < name.size(); i++) {
+    for (int i = 0; i < (int)(name.size()); i++) {
       // I don't trust isalnum() due to locales.  :(
       if ((name[i] < 'a' || 'z' < name[i]) &&
           (name[i] < 'A' || 'Z' < name[i]) &&
@@ -2620,7 +2620,7 @@ void DescriptorBuilder::ValidateSymbolName(
 bool DescriptorBuilder::ValidateQualifiedName(const string& name) {
   bool last_was_period = false;
 
-  for (int i = 0; i < name.size(); i++) {
+  for (int i = 0; i < (int)(name.size()); i++) {
     // I don't trust isalnum() due to locales.  :(
     if (('a' <= name[i] && name[i] <= 'z') ||
         ('A' <= name[i] && name[i] <= 'Z') ||
@@ -2725,10 +2725,10 @@ const FileDescriptor* DescriptorBuilder::BuildFile(
   //   mid-file, but that's pretty ugly, and I'm pretty sure there are
   //   some languages out there that do not allow recursive dependencies
   //   at all.
-  for (int i = 0; i < tables_->pending_files_.size(); i++) {
+  for (int i = 0; i < (int)(tables_->pending_files_.size()); i++) {
     if (tables_->pending_files_[i] == proto.name()) {
       string error_message("File recursively imports itself: ");
-      for (; i < tables_->pending_files_.size(); i++) {
+      for (; i < (int)(tables_->pending_files_.size()); i++) {
         error_message.append(tables_->pending_files_[i]);
         error_message.append(" -> ");
       }
@@ -3029,7 +3029,7 @@ void DescriptorBuilder::BuildFieldOrExtension(const FieldDescriptorProto& proto,
             result->default_value_float_ = numeric_limits<float>::quiet_NaN();
           } else  {
             result->default_value_float_ =
-              NoLocaleStrtod(proto.default_value().c_str(), &end_pos);
+              (float)NoLocaleStrtod(proto.default_value().c_str(), &end_pos);
           }
           break;
         case FieldDescriptor::CPPTYPE_DOUBLE:
@@ -4145,7 +4145,7 @@ bool DescriptorBuilder::OptionInterpreter::SetOptionValue(
                                option_field->full_name() + "\".");
         } else {
           SetInt32(option_field->number(),
-                   uninterpreted_option_->positive_int_value(),
+                   (int32)(uninterpreted_option_->positive_int_value()),
                    option_field->type(), unknown_fields);
         }
       } else if (uninterpreted_option_->has_negative_int_value()) {
@@ -4155,7 +4155,7 @@ bool DescriptorBuilder::OptionInterpreter::SetOptionValue(
                                option_field->full_name() + "\".");
         } else {
           SetInt32(option_field->number(),
-                   uninterpreted_option_->negative_int_value(),
+                   (int32)(uninterpreted_option_->negative_int_value()),
                    option_field->type(), unknown_fields);
         }
       } else {
@@ -4192,7 +4192,7 @@ bool DescriptorBuilder::OptionInterpreter::SetOptionValue(
                                option_field->name() + "\".");
         } else {
           SetUInt32(option_field->number(),
-                    uninterpreted_option_->positive_int_value(),
+                    (uint32)(uninterpreted_option_->positive_int_value()),
                     option_field->type(), unknown_fields);
         }
       } else {
@@ -4215,11 +4215,11 @@ bool DescriptorBuilder::OptionInterpreter::SetOptionValue(
     case FieldDescriptor::CPPTYPE_FLOAT: {
       float value;
       if (uninterpreted_option_->has_double_value()) {
-        value = uninterpreted_option_->double_value();
+        value = (float)(uninterpreted_option_->double_value());
       } else if (uninterpreted_option_->has_positive_int_value()) {
-        value = uninterpreted_option_->positive_int_value();
+        value = (float)(uninterpreted_option_->positive_int_value());
       } else if (uninterpreted_option_->has_negative_int_value()) {
-        value = uninterpreted_option_->negative_int_value();
+        value = (float)(uninterpreted_option_->negative_int_value());
       } else {
         return AddValueError("Value must be number for float option \"" +
                              option_field->full_name() + "\".");
@@ -4234,9 +4234,9 @@ bool DescriptorBuilder::OptionInterpreter::SetOptionValue(
       if (uninterpreted_option_->has_double_value()) {
         value = uninterpreted_option_->double_value();
       } else if (uninterpreted_option_->has_positive_int_value()) {
-        value = uninterpreted_option_->positive_int_value();
+        value = (double)(uninterpreted_option_->positive_int_value());
       } else if (uninterpreted_option_->has_negative_int_value()) {
-        value = uninterpreted_option_->negative_int_value();
+        value = (double)(uninterpreted_option_->negative_int_value());
       } else {
         return AddValueError("Value must be number for double option \"" +
                              option_field->full_name() + "\".");
